Add setup flow.
