+++ crates/core/src/flow/mod.rs
@@ -1,1 +1,2 @@
 pub mod install;
+pub mod setup;
